backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
motor==3.4.0
pymongo==4.7.2
google-generativeai==0.5.4
python-dotenv==1.0.1
pydantic==2.7.1
pydantic[email]==2.7.1
scikit-learn==1.4.2
numpy==1.26.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.0
